+++ tools/setup/setup-wifi-ap.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
+# SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 Stéphane Caron
 # Copyright 2023 Inria
-# SPDX-License-Identifier: Apache-2.0
 #
 # This file incorporates work covered by the following copyright and
 # permission notice:
 #
 #     setup-system.py from github.com:mjbots/quad
 #     Copyright 2018-2020 Josh Pieper
-#     License: Apache-2.0
+#     SPDX-License-Identifier: Apache-2.0
 
 
 """Set up a Raspberry Pi 4b to operate a Wi-Fi access point.