+++ upkie/envs/__init__.py
@@ -1,18 +1,16 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2022 Stéphane Caron
 # SPDX-License-Identifier: Apache-2.0
+# Copyright 2022 Stéphane Caron
 
 import logging
 
 import gymnasium as gym
 
-from .init_randomization import InitRandomization
 from .upkie_base_env import UpkieBaseEnv
 
 __all__ = [
-    "InitRandomization",
     "UpkieBaseEnv",
     "register",
 ]