+++ upkie/envs/wrappers/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2023 Inria
 # SPDX-License-Identifier: Apache-2.0
+# Copyright 2023 Inria
 
 from .add_action_to_observation import AddActionToObservation
 from .differentiate_action import DifferentiateAction