+++ upkie/envs/wrappers/noisify_action.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2023 Inria
 # SPDX-License-Identifier: Apache-2.0
+# Copyright 2023 Inria
 
 import gymnasium
 import numpy as np
+from numpy.typing import NDArray
 
 from upkie.utils.exceptions import UpkieException
 
@@ -16,7 +17,7 @@ class NoisifyAction(gymnasium.ActionWrapper):
     Add noise to the action of an environment.
     """
 
-    def __init__(self, env, noise: np.ndarray):
+    def __init__(self, env, noise: NDArray[float]):
         super().__init__(env)
         if noise.shape != env.action_space.shape:
             raise UpkieException(