+++ upkie/utils/robot_state_randomization.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Copyright 2023 Inria
 # SPDX-License-Identifier: Apache-2.0
+# Copyright 2024 Inria
 
 """!
-Domain randomization of Upkie environments.
+Domain randomization of robot states.
 """
 
 from dataclasses import dataclass
@@ -17,20 +17,20 @@ from scipy.spatial.transform import Rotation as ScipyRotation
 
 
 @dataclass
-class InitRandomization:
+class RobotStateRandomization:
 
     """!
-    Domain randomization parameters for Upkie's initial state.
+    Domain randomization parameters for Upkie's state.
     """
 
     roll: float = 0.0
     pitch: float = 0.0
     x: float = 0.0
     z: float = 0.0
-    v_x: float = 0.0
-    v_z: float = 0.0
     omega_x: float = 0.0
     omega_y: float = 0.0
+    v_x: float = 0.0
+    v_z: float = 0.0
 
     def update(
         self,
@@ -38,11 +38,31 @@ class InitRandomization:
         pitch: Optional[float] = None,
         x: Optional[float] = None,
         z: Optional[float] = None,
-        v_x: Optional[float] = None,
-        v_z: Optional[float] = None,
         omega_x: Optional[float] = None,
         omega_y: Optional[float] = None,
+        v_x: Optional[float] = None,
+        v_z: Optional[float] = None,
     ) -> None:
+        """!
+        Update some fields.
+
+        @param roll Roll angle for the rotation from the randomized frame to
+            the parent frame (Euler Z-Y-X convention).
+        @param pitch Pitch angle for the rotation from the randomized frame to
+            the parent frame (Euler Z-Y-X convention).
+        @param x Translation of the randomized frame along the x-axis of the
+            parent frame.
+        @param z Translation of the randomized frame along the z-axis of the
+            parent frame.
+        @param omega_x Angular velocity from the randomized frame to the parent
+            frame, expressed in the randomized frame, along the x-axis.
+        @param omega_y Angular velocity from the randomized frame to the parent
+            frame, expressed in the randomized frame, along the y-axis.
+        @param v_x Linear velocity from the randomized frame to the parent
+            frame, expressed in the parent frame and along the x-axis.
+        @param v_z Linear velocity from the randomized frame to the parent
+            frame, expressed in the parent frame and along the z-axis.
+        """
         if roll is not None:
             self.roll = roll
         if pitch is not None:
@@ -70,17 +90,9 @@ class InitRandomization:
         return ScipyRotation.from_euler("ZYX", yaw_pitch_roll)
 
     def sample_position(self, np_random) -> NDArray[float]:
-        default_position = np.array([0.0, 0.0, 0.6])
-        return default_position + np_random.uniform(
+        return np_random.uniform(
             low=np.array([-self.x, 0.0, 0.0]),
             high=np.array([+self.x, 0.0, self.z]),
-            size=3,
-        )
-
-    def sample_linear_velocity(self, np_random) -> NDArray[float]:
-        return np_random.uniform(
-            low=np.array([-self.v_x, 0.0, -self.v_z]),
-            high=np.array([+self.v_x, 0.0, +self.v_z]),
             size=3,
         )
 
@@ -88,5 +100,12 @@ class InitRandomization:
         return np_random.uniform(
             low=np.array([-self.omega_x, -self.omega_y, 0.0]),
             high=np.array([+self.omega_x, +self.omega_y, 0.0]),
+            size=3,
+        )
+
+    def sample_linear_velocity(self, np_random) -> NDArray[float]:
+        return np_random.uniform(
+            low=np.array([-self.v_x, 0.0, -self.v_z]),
+            high=np.array([+self.v_x, 0.0, +self.v_z]),
             size=3,
         )